theHarvester-4.0.3.ebuild

+++ net-analyzer/theHarvester/theHarvester-4.0.3.ebuild
@@ -14,8 +14,8 @@ SRC_URI="https://github.com/laramies/theHarvester/archive/${PV}.tar.gz -> ${P}.t
 LICENSE="GPL-2"
 SLOT="0"
 #dev-python/plotly not stable, no x86 for orjson
-KEYWORDS="~amd64 ~arm64"
-# IUSE="test"
+KEYWORDS="~amd64"
+IUSE="spyse"
 REQUIRED_USE="${PYTHON_REQUIRED_USE}"
 
 # RESTRICT="!test? ( test )"
@@ -29,21 +29,22 @@ RDEPEND="${PYTHON_DEPS}
 	dev-python/beautifulsoup4[${PYTHON_USEDEP}]
 	dev-python/censys[${PYTHON_USEDEP}]
 	dev-python/certifi[${PYTHON_USEDEP}]
-	>=dev-python/dnspython-1.16.0[${PYTHON_USEDEP}]
+	>=dev-python/dnspython-2.1.0[${PYTHON_USEDEP}]
 	dev-python/fastapi[${PYTHON_USEDEP}]
+	>=dev-python/lxml-4.6.4[${PYTHON_USEDEP}]
 	dev-python/netaddr[${PYTHON_USEDEP}]
-	dev-python/orjson-bin
+	dev-python/ujson[${PYTHON_USEDEP}]
 	dev-python/pyppeteer[${PYTHON_USEDEP}]
-	>=dev-python/pyyaml-5.4.1[${PYTHON_USEDEP}]
+	>=dev-python/pyyaml-6.0[${PYTHON_USEDEP}]
 	dev-python/requests[${PYTHON_USEDEP}]
 	dev-python/retrying[${PYTHON_USEDEP}]
 	>=dev-python/shodan-1.25.0[${PYTHON_USEDEP}]
 	dev-python/slowapi[${PYTHON_USEDEP}]
 	dev-python/starlette[${PYTHON_USEDEP}]
-	dev-python/lxml[${PYTHON_USEDEP}]
-	dev-util/unicorn[${PYTHON_USEDEP}]
+	dev-python/uvicorn[${PYTHON_USEDEP}]
 	dev-python/uvloop[${PYTHON_USEDEP}]
 	"
+#	dev-python/spyse-python[${PYTHON_USEDEP}]
 
 DEPEND="${RDEPEND}"
 	# test? (
@@ -52,7 +53,14 @@ DEPEND="${RDEPEND}"
 	# )"
 
 # distutils_enable_tests pytest
+
 src_prepare() {
+	if use spyse; then
+		eerror "FIXME: spyse-python"
+		eerror "https://github.com/spyse-com/spyse-python/issues/7"
+		exit
+	fi
+
 	# network required for tests
 	rm -r tests || die
